+++ src/plugins/reporter-mantisbt.c
@@ -463,7 +463,7 @@ int main(int argc, char **argv)
         exit(0);
     }
 
-     int bug_id = 0;
+     unsigned long bug_id = 0;
 
     /* If REMOTE_RESULT contains "DUPLICATE 12345", we consider it a dup of 12345
      * and won't search on MantisBT server.
